requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.5
djangorestframework==3.12.2
flashgeotext==0.2.0
flashtext==2.7
joblib==1.0.0
loguru==0.4.1
nltk==3.5
numpy==1.19.5
pydantic==1.4
pytz==2020.5
regex==2020.11.13
scikit-learn==0.24.0
scipy==1.6.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.55.2
